move assertETHSanityChecks to fixtures file
